Finished generateInsert style question generator
The difficulty selects the number of columns. The generator then selects a table with at least that many columns and removes them until there are a sufficient amount remaining. It then generates random data of the appropriate data type to be inserted.

Also added another function to the textDatabaseHandler that returns the CURRENT schema, which will reflect any removed columns