Completed generateDelete, except for hard difficulty
hard difficulty requires the completion of the query-style questions.

The approach for DELETE is to select a database whose size depends on the difficulty. On easy there can be no foreign keys in the database, and on hard it will require a subquery. It then generates a bunch of random data and selects on entry of that data to be deleted.