Completed UPDATE-style question generation
Likewise to DELETE, hard difficulty is not yet implemented since it requires queries

Works the same as DELETE-style where it selects a database of a specific size then selects a column to be affected. Where it differs is that it generates a new random value to updated rather than an existing value.

Another difference is that it only uses a conditional (WHERE $col = $val) is medium or hard questions.