Finished basic QUERY-style autogen
The approach is difficulty selects the number of joins and the amount of clauses (such as ORDER BY). It then selects a database of appropriate size and with an appropriate amount of foreign keys and fits them together some way that will satisfy the constraints imposed by the difficulty.

It is not finished! It does not add any clauses yet or conditions (other than the joins' conditions)