Added the ability for random tables to reference others
The reference doesn't hold the correct functionality for generating the referenced databases, but the primary looks correct. Still a few changes need to be made, namely clauses.

Removed '%' as a placeholder in columnNames.txt since foreign key generation would conflict.

Added a list of possible table names as a text file, similar to column names