Refactored noisy data generation
The goal for refactoring was to accomplish two things: allow for generated data to be unique, and reduce the number of file reads.

To accomplish the first, it (when possible) pops values from a unique list, otherwise it keeps generating values until one is unique.

The second was done by changing the generation method from row-by-row to column-by-column. This required refactoring the loadAllNoisyData() function, which is good anyways because it's far cleaner now